+++ app/src/main/java/com/cysj/zziririt/presentation/postcontent/CommentItem.kt
@@ -1,0 +1,7 @@
+package com.cysj.zziririt.presentation.postcontent
+
+data class CommentItem(
+    val nickname: String,
+    val time: String,
+    val text: String,
+)
