+++ app/src/main/res/drawable/ic_note_zziririt.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="64dp"
+    android:height="64dp"
+    android:viewportWidth="64"
+    android:viewportHeight="64">
+  <path
+      android:pathData="M32,32m-32,0a32,32 0,1 1,64 0a32,32 0,1 1,-64 0"
+      android:fillColor="#535353"/>
+  <path
+      android:pathData="M42.606,37.067H50.776C50.686,37.175 50.597,37.27 50.508,37.351C50.418,37.432 50.338,37.5 50.267,37.554L50.186,37.662L33.499,53.912C33.178,54.237 32.785,54.4 32.32,54.4C31.856,54.4 31.463,54.237 31.142,53.912L14.427,37.608C14.338,37.572 14.151,37.391 13.865,37.066H23.749C24.142,37.066 24.494,36.945 24.807,36.701C25.119,36.457 25.32,36.146 25.41,35.766L27.285,28.156L32.374,46.221C32.481,46.582 32.686,46.88 32.99,47.114C33.293,47.349 33.642,47.466 34.035,47.466C34.41,47.466 34.749,47.349 35.053,47.114C35.356,46.88 35.561,46.582 35.668,46.221L39.579,33.085L41.079,36.118C41.401,36.75 41.91,37.067 42.606,37.067ZM56.32,25.475C56.32,28.093 55.401,30.801 53.561,33.6H43.678L40.704,27.614C40.561,27.308 40.334,27.064 40.021,26.883C39.709,26.703 39.383,26.63 39.043,26.667C38.24,26.757 37.74,27.172 37.543,27.912L34.088,39.558L28.838,20.979C28.731,20.618 28.521,20.32 28.209,20.085C27.896,19.851 27.544,19.733 27.151,19.733C26.758,19.733 26.41,19.855 26.106,20.099C25.802,20.343 25.606,20.654 25.517,21.033L22.41,33.6H11.079C9.24,30.801 8.32,28.093 8.32,25.475C8.32,21.503 9.454,18.397 11.722,16.158C13.99,13.919 17.124,12.8 21.124,12.8C22.231,12.8 23.361,12.994 24.512,13.382C25.664,13.77 26.736,14.294 27.727,14.953C28.718,15.612 29.57,16.23 30.285,16.808C30.999,17.386 31.677,18 32.32,18.65C32.963,18 33.642,17.386 34.356,16.808C35.07,16.23 35.923,15.612 36.914,14.953C37.905,14.294 38.977,13.77 40.128,13.382C41.28,12.994 42.41,12.8 43.517,12.8C47.517,12.8 50.651,13.919 52.918,16.158C55.186,18.397 56.32,21.503 56.32,25.475Z"
+      android:fillColor="#00FFA3"/>
+</vector>
